Use strings instead of integers for notification actions internally
The backend typically natively use strings for action ids and we convert those to indices

By using strings ourselves we match that better and gain more flexibility

+++ src/notifybyportal.cpp
@@ -200,7 +200,7 @@ void NotifyByPortal::onPortalNotificationActionInvoked(const QString &id, const 
 
     KNotification *n = *iter;
     if (n) {
-        Q_EMIT actionInvoked(n->id(), action.toUInt());
+        Q_EMIT actionInvoked(n->id(), action);
     } else {
         d->portalNotifications.erase(iter);
     }
